requirements.txt
# Streamlit App
streamlit>=1.29.0

# PDF Parsing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment variables
python-dotenv>=1.0.0

# Email notifications
schedule>=1.2.0
